fix(nslkdd): make sure we pass numpy arrays to fit
We used to pass panda series as labels in the model.fit() method, and it
used to work because the underlying storage was somehow compatible with
what we expect.
This doesn't work anymore since recent updates in the panda , sklearn and
 imblearn packages.
We now explicitly convert labels series to numpy array.

Signed-off-by: David Corvoysier <dcorvoysier@brainchip.com>