refactor(examples/plot_ds_cnn_cifar10): rework of the example
The example needed a rework to include latest changes in CNN2SNN and
akida models (quantize API and CLI). The VGG part of the example has been
removed to ease reading and understanding, the model is still available in the
API reference/Akida models/Model zoo section of the documentation.