feat: add links to source code for functions and classes
sphinx.ext_viewcode extension looks at documented packages and automatically
find the associated source code. A separate HTML page will be output for each
module with a highlighted version of the source code, and a link will be added
to all object descriptions that leads to the source code of the described
object. A link back from the source to the description will also be inserted.