fix(examples): adjust get_statistics usage
The statistics are now accumulated directly by the akida runtime and a snapshot
can be retrieved on-demand with get_statistics(). The snapshot is NOT updated on
subsequent forward calls, and get_statistics must be called again to get updated
values.

This is a change from the previous usage where one needed to instantiate a
Statistics object that gets updated every call to forward.