feat(advanced_cnn2snn): add section on very high scale factors
A section in the advanced cnn2snn tutorial is added at the end to give info
on how to deal with too high scale factors (that leads to a Runtime Error at
Akida inference time).
The proposed solution is to add constraint on BN gammas and retrain to limit
too small gammas in BN layers.