chore(docker): install explicitly onnx and onnruntime
Although this is QuantizeML's job, optimum forces the installation
of these packages on higher versions. That is why we install and
set the version of them to prevent downgrading later
when installing QuantizeML.

Note that downgrade WILL ALWAYS HAPPEN for CI nightly anyway :
the protobuf version of this image (3.20.3) is not compatible with
tensorflow-datasets==4.9.0 and tensorflow-metadata==1.13.0.
That is why in this pipeline protobuf is forced to 3.19.6,
which will force a downgrade of ONNX when quantizeml=0.7.3
is installed to 1.12.0.