chore(drone): rework onnx examples
Apparently there are conflicts between the TF2.10 docker image
NVCC version and pytorch when it uses the GPU. Since it is not
the purpose of the example to work on GPU, we decided to turn it
to CPU. This has no impact on the performance of the example:
"(...) Why don't I notice MASSIVE speedup compared to CPU?
Because your network is really small" (extracted from the
pytorch base example).

Additionally, we have decided to remove the dependency on the
ONNX opset version, since QuantizeML 0.7.2 the conversion is
implicit to the one required for quantizing. Experimentally this allows
working with a wider range of pytorch versions.